helpers: split Image and ImageUtils
